refactor to a single util function and remove ensurePathLoaded and ensureAncestorsLoaded
