+++ packages/@uppy/provider-views/src/ProviderView/ProviderView.tsx
@@ -25,6 +25,7 @@ import getBreadcrumbs from '../utils/PartialTreeUtils/getBreadcrumbs.js'
 import getCheckedFilesWithPaths from '../utils/PartialTreeUtils/getCheckedFilesWithPaths.js'
 import getNumberOfSelectedFiles from '../utils/PartialTreeUtils/getNumberOfSelectedFiles.js'
 import PartialTreeUtils from '../utils/PartialTreeUtils/index.js'
+import { materializePath } from '../utils/PartialTreeUtils/pathLoaderCore.js'
 import shouldHandleScroll from '../utils/shouldHandleScroll.js'
 import AuthView from './AuthView.js'
 import Header from './Header.js'
@@ -458,13 +459,13 @@ export default class ProviderView<M extends Meta, B extends Body> {
           return { items, nextPagePath }
         }
 
-        const { partialTree: materializedTree, targetId } =
-          await PartialTreeUtils.ensurePathLoaded(
-            partialTree,
-            folderId,
-            apiList,
-            this.validateSingleFile,
-          )
+        const { partialTree: materializedTree, targetId } = await materializePath(
+          partialTree,
+          folderId,
+          apiList,
+          this.validateSingleFile,
+          { includeTargetFirstPage: true },
+        )
 
         // Clean search nodes and navigate to the real folder id
         const cleanedTree = this.#removeSearchNodes(materializedTree)
@@ -793,13 +794,13 @@ export default class ProviderView<M extends Meta, B extends Body> {
           return { items, nextPagePath }
         }
         // 1) Ensure ancestors and target node exist in real tree
-        const { partialTree: withAncestors, targetId } =
-          await PartialTreeUtils.ensureAncestorsLoaded(
-            partialTree,
-            rawId,
-            apiList,
-            this.validateSingleFile,
-          )
+        const { partialTree: withAncestors, targetId } = await materializePath(
+          partialTree,
+          rawId,
+          apiList,
+          this.validateSingleFile,
+          { includeTargetFirstPage: false },
+        )
         // 2) Toggle the real node
         const realId = (targetId ?? '').toString()
         const toggledTree = PartialTreeUtils.afterToggleCheckbox(
