+++ packages/@uppy/provider-views/src/utils/PartialTreeUtils/index.ts
@@ -2,14 +2,11 @@ import afterFill from './afterFill.js'
 import afterOpenFolder from './afterOpenFolder.js'
 import afterScrollFolder from './afterScrollFolder.js'
 import afterToggleCheckbox from './afterToggleCheckbox.js'
-import ensureAncestorsLoaded from './ensureAncestorsLoaded.js'
-import ensurePathLoaded from './ensurePathLoaded.js'
+
 
 export default {
   afterOpenFolder,
   afterScrollFolder,
   afterToggleCheckbox,
   afterFill,
-  ensurePathLoaded,
-  ensureAncestorsLoaded,
 }
